requirements.txt
fastapi
uvicorn[standard]
pydantic

# code-cartographer (Phase A.5) — universal AST parsing for Python/JS/Kotlin
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-kotlin>=0.3.0

# code-cartographer — annotation files frontmatter (YAML)
PyYAML>=6.0

# code-cartographer — coverage map (Phase A.7) symbol/range ↔ tests linking
pytest-cov>=5.0
coverage>=7.0

# V2.1 Postgres migration — SQLAlchemy 2.x ORM + Alembic + UUID v7
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
uuid_utils>=0.10

# V2.1 Postgres tests — testcontainers spin un PG 16 jetable par session pytest
testcontainers[postgres]>=4.0

# V2.2 — chiffrement at-rest AES-256-GCM applicatif (champs Art.9 RGPD)
cryptography>=42.0

# V2.0.5 — observability foundation (logs JSON structurés, request_id corrélation)
structlog>=24.1

# V2.3 — auth foundation (argon2id password hashing + JWT HS256 access/refresh)
argon2-cffi>=23.1
PyJWT[crypto]>=2.8.0

# V2.3.2 — Google OAuth (httpx for token/JWKS fetch, PyJWT[crypto] for ID token RS256)
httpx>=0.27

# V2.3.3.1 — global rate-limiting (slowapi = FastAPI port of Flask-Limiter)
slowapi>=0.1.9,<1.0
